+++ admin/frontend/pages/roles/create.js
@@ -22,13 +22,10 @@ class CreateRole extends Component {
   }
 
   async load () {
-    const body = await api.get(
-      '/admin/roles',
-      {
-        start: 0,
-        limit: this.cursor.get('pageLength') || 10
-      }
-    )
+    const body = await api.get('/admin/roles', {
+      start: 0,
+      limit: this.cursor.get('pageLength') || 10
+    })
 
     this.cursor.set({
       page: 1,
@@ -36,6 +33,7 @@ class CreateRole extends Component {
       items: body.data,
       pageLength: this.cursor.get('pageLength') || 10
     })
+
     this.context.tree.commit()
   }