+++ models/role.js
@@ -13,6 +13,8 @@ const roleSchema = new Schema({
   dateCreated: { type: Date, default: moment.utc },
   uuid: { type: String, default: v4 },
   isDeleted: { type: Boolean, default: false }
+}, {
+  usePushEach: true
 })
 
 roleSchema.methods.toPublic = function () {