+++ models/user.js
@@ -28,6 +28,8 @@ const userSchema = new Schema({
 
   uuid: { type: String, default: v4 },
   apiToken: { type: String, default: v4 }
+}, {
+  usePushEach: true
 })
 
 userSchema.pre('save', function (next) {