+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Query params to make filter functionality easier on list routers
+- Added csv exports to ListPageComponents when `exportFormatter` function is present
 
 ### Fixed
 - Mongoose models need `usePushEach: true` on all models with array attributes to prevent `Unknown modifier: $pushAll` error